Implemented MDL pitch envelopes
MDL envelopes are all wrong. This commit fixes the general structure
but there's still a lot to fix in timing, we need a couple of MDL test
modules to adjust this. Digitrakker runs fine in dosbox so creating
test modules should be easy.

Also updated README.trackers with Digitrakker quirks.