Updating Audacious plugin to work with Audacious 1.4.0
Audacious is changing the API to remove XMMS legacy, new operations
are already in place but the old functions are still there, so you
don't know which ones work and which ones don't. E.g. get_song_info
is still there, but it doesn't work, get_song_tuple should be used
instead.