fix terminal handling in Cygwin
from	daniel åkerud <daniel.akerud@gmail.com>
to	xmp-devel@lists.sourceforge.net
date	Tue, Jul 28, 2009 at 9:59 AM
subject	[xmp] [PATCH] Fix blocking read() in process_echoback for Cygwin

Under Cygwin, the read() in process_echoback blocks because VTIME = 0
is not handled correctly. To fix this you can either:

1. Enable "tty emulation" in Cygwin using an environment variable.
http://www.mail-archive.com/cygwin@cygwin.com/msg99417.html
For me this is _very_ slow and I can see the characters as they are
typed out when running xmp. I have not investigated why this is
happening, but there is of course a reason why this mode is not
enabled by default.

2. Do a select() before read()ing if the platform is Cygwin.
This makes Cygwin builds work out of the box with no fiddling around,
but does impose a neglectible cpu overhead (for Cygwin builds only).

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>