[Android] Change task layout
Previous version had the Player activity as singleInstance, making
it a root activity in a task and preventing the task switcher to
switch to the playlist menu. Now the playlist menu is set as singleTask
and the Player set as singleTop, handling new intents as they arrive.

See http://developer.android.com/guide/topics/manifest/activity-element.html

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>