Fix pattern delay and break event retrieval
The Welder reports that Condom Corruption is played incorrectly after
a pattern break (D31) that should actually land on line 32 of the next
pattern.

Investigation has shown that many trackers (Protracker 3.15, FT2, ST3,
but not Milkytracker) ignore target row events if pattern break and
pattern delay are issued in the same line. Added a special flag to
handle this situation instead of reorganizing internals because we
want this behavior to be switchable via quirks if needed.

Note: Condom corruption still reports 3:27 instead of 3:29. Before this
fix it reported 3:25 instead of 3:27, so this error has a different origin.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>