Better fix for sample mixing
S3M and IT use the sample end parameter as the first sample after the loop,
we use it as the last sample of the loop. Duplicate the last sample if
loop end is after the sample end. Also prevent memory access violations
by correctly setting the sample loop start when mixing.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>