[test] Add decompression tests
Add depacker tests using custom framework similar to check. The reason
for not using an existing framework is that we want to run tests on strange
platforms where probably the existing test framework won't run or will be
a pain to port.

Also move existing tests out of the way while we decide what to do with it.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>