Fix sudden muting of virtual channel
Manual volume setting for virtual channels should follow the volume of
their root channels, otherwise a virtual channel may have its volume
set to 0 (when its virtual channel number is larger than 64).

Along with the increase of the maximum number of mixer voices from 64 to
128, this change fixes replay of the arpeggiator using virtual channels
in ambio.it position 4, channels 9-12.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>