Fix global volume setting in module scan
Global volumes are collected for each position in scan phase to ensure
that replay will be performed with the correct volume when starting a
module from a given position. Global volume was being clamped at 64
instead of the maximum global volume for that format, m->gvolbase. Change
hardcoded values to the appropriate global volume base.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>