Update Mac OS X dylib compatibility version
I hope this is being done correctly, as online documentation seems to
be scarce. Check http://stackoverflow.com/questions/15905310:

The version number checks that dyld performs are limited to ensuring that
the compatibility version of the library being loaded is higher than the
compatibility version of the library that was used at build time. The
current version of the library is something that's programmatically
accessible, but isn't used by dyld when evaluating whether it should load
a particular library.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>