Add mechanism to skip sample loading
Winamp, Audacious and other plugins may want to know the duration of
a module without playing it. We can load a module in a different context
and scan it to get the duration. Being able to do it without loading
all the module samples is faster, so add a mechanism to skip sample loading
for cases where we only want to scan the module patterns.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>