[test] Add envelope reset after portamento tests
When a tone portamento effect is performed, we expect it to always
reset the envelope if the previous envelope already ended. If it
still didn't end, reset envelope in XM (and IT compatible GXX mode)
but not in standard IT mode.

Update expected results from existing tests to reflect this fix.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>