Fix envelope reset on portamento effect
When a tone portamento effect is performed, we expect it to always
reset the envelope if the previous envelope already ended. If it
still didn't end, reset envelope in XM (and IT compatible GXX mode)
but not in standard IT mode.

See olympic.xm and portamento_envelope_reset tests. Reported by
ArtRemix <leartremix@gmail.com>.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>