Add simple library test
To check whether the library was built properly we'll load and play a
test module. This should make sure that a depacker, a loader, the player,
the mixer and a portion of the API are working correctly. To check for
regressions a set of comprehensive tests (the old test suite) is executed
by the developer before publishing the tarball.

This also addresses issues related to shipping binary data with unclear
licensing terms in the distribution package.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>