Handle MED hold/decay event
MMD0/1/2/3 MED decay events are set as instrument numbers without notes.
The instrument note is the decay value for regular instruments or jumps
to the volume table for synth instruments.

This patch sets an FX_MED_DECAY effect for decay events, but actual
decay or JVS effect processing is left to a future patch.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>