Relicense ptpopen as public domain
Date: Thu, 29 Aug 2013 08:23:36 +0200
From: Kurt Keller <kk@parity-software.com>
To: Claudio Matsuoka <cmatsuoka@gmail.com>

Hi Claudio,

I am surprised how long this piece of code hangs around. I wrote it in the
somewhere in the 90ies.

Yes you can assume, that this piece of code is in the public domain. I do
not claim any rights on it. Do whatever you want to do with it and I hope
it will be still useful.

Thank you for asking!

Regards
Kurt.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>