Refactor common code from module loaders
The module loading code in  functions xmp_load_module() and
xmp_load_module_from_memory() were essentially the same, extract them to
a common function.

Additionally, don't set the module name to the file name if the name is
empty, it's up to the application to decide what to do with empty names.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>