Add ZIP I/O capabilities to stdio wrappers
The main purpose of this refactoring is to allow the player to load
zip files without needing a temporary file, and to allow access to
a specific file inside the zip archive. This patch only deals with
the first issue -- dealing with individual files inside the zip
archive is left to a forthcoming commit.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>