Handle new instrument in MED extras play call
Instead of passing cooked flags for new note and instrument to the extras
play call, just pass the tick number and use it with the raw flags we're
already passing. This allows us to check for any new flags, such as the
new instrument flag, and we'll use this to not reset the MED synth waveform
when a note is used without the instrument number (see MED.HC Under Control).

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>