Fix MOD note range enforcing
Jason Gibson reports an off-key note in Jason Thomale's Apathy.mod. This
module was incorrectly detected as a Noisetracker module, even if a note
outside of the Noise/Sound/Protracker range was used. In this case, don't
detect as an Amiga tracker (use unknown tracker instead) an don't enforce
amiga limits.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>