Fix temporary file creation in Windows
The GNU mkstemp() implementation was replaced with a BSD version somewhere
between 4.3.0 and 4.3.1 and the new implementation didn't set O_BINARY
when opening the temporary file. In Windows, it tried to translate CR-LF
sequences when depacking prowizard files and created invalid modules.

This bug was reported by Andreas Argirakis and Eric Lévesque.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>