Fix scan of IT pattern row delay
Vincent Voois' "The Snarfs" stopped in the first row because there's a
pattern row delay effect (SE1) in channel 31 and we play extra empty
rows before the actual events. When that pattern/row is also the module
loop point, xmp thinks the module looped over and stops replay. To fix
this, add the extra rows to the scan counter.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>