Refactor module scan frame counter
The scan code is one of the oldest and ugliest parts of the player. It uses
obscure logic to gather timing and volume data from the module prior to
actual replay. We'll try to rename and add meaning to some of the variables
in an attempt to make this code easier to maintain.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>