Replace invalid patterns in order list
Invalid patterns in order list, loaded from a corrupt or invalid file may
cause a segfault. Replace them with a valid pattern (or we could also
reject the module entirely because it's corrupt). We replaced the invalid
pattern to be more tolerant, and because Schism Tracker seems to do that
as well.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>