Fix instrument volume in Amiga split channels
Dennis Lindroos tells that: "the big thing is that instrument volumes are
not used on split channels, i.e. instrument volume is ignored in such
channels.

Split voices always at the volyme of the hardware channels because the
software mixing of the samples in split channels always happens at full
(7 bits + 7 bits = 8 bits sample) volume.

So, on a split hardware channel, the volume is only set by the V (0x1f)
command, not by the sample/instrument, for example these two split channels:

(V is command 0x1F, initialized at 0x40)

C-4 V4F  --- 000    <- the volume is sliding down
--- V4F  --- 000
C-3 000  --- 000    <- this note is not heard anymore
--- 000  D-3 000    <- ditto
--- 000  E-3 000    <-
C-4 V40  --- 000    <- this is now heard (and also the E-3)
--- 000  G-3 000"

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>