Allow invalid 0x2x finetune when testing for MOD
Chris Spiegel says:

(https://github.com/cmatsuoka/libxmp/issues/48)

sandman.mod from here is not recognized by libxmp, due to the following checks in src/loaders/mod_load.c, which are done as a “[s]anity check to prevent loading NoiseRunner and other module formats with valid magic at offset 1080”:

if (hio_read8(f) & 0xf0)                /* test finetune */
        return -1;
if (hio_read8(f) > 0x40)                /* test volume */
        return -1;

If I modify the above to not return failure, the file loads and plays fine. Apparently this module looks enough like a NoiseRunner module to trigger the tests, but it's not one. I'm not sure if the module is partially corrupted or if the test result is a false positive.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>