Add flags to configure player personality
Allow the user to override tracker detection and set the player personality
to play a module as a generic MOD player, Noisetracker, Protracker, generic
S3M player, Scream Tracker 3, Scream Tracker 3 with GUS, generic XM player,
Fast Tracker 2, Impulse Tracker, or Impulse Tracker in sample mode.

This is intended to be used in cases where autodetection is impossible and
we can't guess what the original author wanted to do.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>