Use Sound Blaster mode as default in S3M
Part of the experiment of player personality flags. ST3+GUS behavior can be
enabled using XMP_FLAGS_ST3GUS, so we can leave the default as SB since
there are modules relying on this.

See Saga Musix comments on commit #7795a8e in GitHub:

"Claudio, I'm not sure if it's triggered by the same quirk in xmp (in OpenMPT
 it is), but now I'm not sure if using the GUS behaviour is really the way to
 go for ST3-made S3Ms either.

 http://modarchive.org/module.php?169007

 Go to order 9 / pattern 7, row 48 (around 1:17 into the tune).
 On row 54, there is a sample change with portamento. With SoundBlaster, ST3
 switches to the new sample (as intended and like IT does it too), but GUS
 keeps playing the old sample (and applies the portamento). I am now
 contemplating to always apply the SB behaviour by default (but optionally
 let people enable GUS behaviour for files that need it), since so far I have
 not found any files that rely on the GUS behaviour."

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>