Fix MOD loader period to note conversion
Leilei reports in https://github.com/cmatsuoka/libxmp/issues/59:

"High notes seem to get bumped up one note for some reason, like F#8
 becoming G-8.

Replace a complicated table lookup with a simple logarithmic operation to
directly obtain the note value. We tried to avoid floating point operations
back in 1996 but it doesn't make sense anymore.

Note: we're using log10()/M_LN2 instead of log2() because it's not supported
in all platforms we want to build on. I can live with that.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>