Fix XM note release after envelope position set
Lionel Debroux reports:

"in the weirdness department, there's farewell_tear.xm . ocp clearly
 gets the first part wrong, as some notes rendered by both mplayer and
 xmp are missing, and the ocp output sounds weird. However, the second
 part sounds better with both ocp and mplayer than it does with xmp.
 It's entirely possible that the tracker used to produce that file was
 buggy, and libxmp is in fact right; however, differences wrt. other
 renderers are noteworthy."

It turned out that this module uses the "set envelope position" XM effect
in the second part, setting the envelope to a position after the sustain
point. Xmp was then returning the envelope to the sustain point, making
the staccato organ sound to be sustained. Now we check if we're after the
sustain point (or sustain loop end point), and release the note to avoid
returning to the sustain point.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>