Remove Adlib support
Adlib suport was very broken, mostly because channels were not being rendered
separately. I fixed RAD support to be more or less usable, but overall support
wasn't perfect so I decided to focus on quality of the main formats instead of
keep hammering on badly supported features. Also it removes an LGPLed component
from the system. If we replace inflate and other minor components with versions
implemented under a more permissive license, we can have a fully MIT-licensed
library.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>