Fix IT envelope reset on tone portamento
Lionel Debroux reports:

"On saga_musix_-_suburban_streets.it, libxmp gets the duration right,
 while libmodplug underestimates it by over 1'30"; but neither library's
 rendering of that track matches the OpenMPT rendering provided by Saga
 Musix as an OGG file.
 For instance, both libxmp and libmodplug render a high-pitched, slightly
 tremolo, and off-tune note from ~1'30" to ~2'17". libmodplug uses too
 few voices to render the track and the note even occasionally
 disappears. But the out-of-place-sounding note is not in the OGG file."

It happens that the rules to reset envelope are a bit more complex than what
we were using: it happens when we set the instrument, or when there's no tone
portamento. This should cover test_player_it_portamento_after_keyoff,
Suburban Streets, and not break other regression tests.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>