Fix mixer loop sample offset bug
This is an ages old bug that was spotted after testing IT sustain loop
with Saga Musix's atomic_playboy.it. The extra sample should be added
only when looping, and not in the general case. This breaks a lot of the
existing regression tests, they'll be updated in the next commits.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>