Fix global volume setting on module loop
Travis Evans reports:

"I'm not sure whether this is technically a bug, but I discovered a
 module with a strange behavior:  When played by xmp, the play counter
 advances, but I get no sound unless I perform a manual seek to somewhere
 in the song.

 The module is at:
 http://lite.modarchive.org/index.php?request=view_by_moduleid&query=44945

 With recently-released xmp 4.1.0/libxmp 4.4.0, it works if the "-e xm"
 option is used.  There's a weird order at the beginning of the song that
 seems to serve no purpose other than maybe to purposely exploit FT2's
 E60 bug or something."

It seems that this problem is more complex and related to how a position
is reused at a different row entry point. Let's see if this fixes this
issue, otherwise we'll have to add position info per row.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>