Use block reading in XM instrument loader
This is something that should be used in all loaders in order to
correctly check if loaded values are valid, but was postponed
because it was boring. I guess it's time to start fixing things.

This should address truncated files in a generic way, including
Boobiesqueezer and XMLiTE stripped XMs (thanks, Saga Musix!)

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>