Handle IT fadeout NNA when instrument doesn't fade
Not sure if this is the correct fix. Problem in azo_cotb.it reported
by Lionel Debroux is caused by an instrument with fade NNA but fadeout
value 0 and an envelope with sustain loop. In this case, release the
instrument from the sustain loop to prevent notes to being held on
forever.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>