Simplify fadeout logic
The song will fade out only after we see that it's continuing past the
play_length predicted by the file metadata. The UI will display only up 
to play_length.  

We track the elapsed time ourselves and properly account for tempo 
changes. This is because game-music-emu keeps track of time always 
assuming a 1.0x playback speed.