Fix voice muting lost state (#83).
Rename voices to voiceMask. It's an array of booleans.

Error originated in handleSequencerStateUpdated() because the App was resetting all voices to enabled on every state update.

I could have removed that line, but then the voice mask would be stale when the song changes. It's more of a problem to make the App state a source of truth about the player. Now the players are source of truth and they have to implement a getVoiceMask() method.

The pattern for now is to copy state from the player into React state whenever the player tells it to.