Fix playback speed (tempo) state (#83)
Each player should be the source of truth about the playback speed, since each player might have different limitations or idiosyncrasies that influence the speed.

Ideally each player engine will expose a "get playback speed" method, but in the meantime the speed can be stored in the JS player adapters.