Hoist playContexts from Browse component to App state
Renamed from contexts to playContexts.

This was done in order to activate the correct context from a ?play=... param. The context in the sequencer has to match the context in the Browse component by referential equality in order to highlight the currently playing track.

There is no need for play contexts to be confined to Browse component. They are shared with the sequencer and derived directly from the directory listings, which are already stored in App state.

Also, do not share entire directory map and entire context map with Browse and BrowseList. They only need to know about the currently displayed listing/playContext.