Fix play context handling in search results
Search results don't retain referential equality with play context across tab navigations, because a new array is created when the Search component is mounted.

Technically, this is a reasonable behavior, because the search results might be different next time the component is mounted.

One solution (used here) is to relax a bit and not check that the results and context are pointing to the same array. Instead, check that the array *items* are the same at current play index. Of course, this can be wrong at some edge cases.

The more correct solution would be to cache the search results (with RequestCache) so that the search results are idempotent (within a session).