Remove directory list sorting logic from App.js. Filter playContext to files only
This was a long time coming, since the client was doing a lot of redundant work.

We also do this to simplify work of generating the playContext lists, since now we just filter out items where type=='file' to create those contexts. The idx property now comes from the browse endpoint.