Fix end of MIDI silence detection
`buffer` is a number (Emscripten pointer), and `buffer[i]` wasn't throwing an error.

It might be time to switch to TypeScript.

The code is janky, since now passing Emscripten runtime method (getValue) into MIDIFilePlayer, but at least it actually works. End of softsynth MIDI is no longer an abrupt transition.