Add ListView keyboard interaction
Upgrade Favorites tab to a virtualized list.
Replace *react-virtual-list* with *react-virtualized*.
Create VirtualizedList.js (BrowseList.js is no longer needed).
And add Winamp-style "list view" affordances. This includes:

* click to select a row (the "cursor")
* press arrow keys to move cursor up/down
* double click to play/navigate cursor
* press enter to play/navigate cursor
* restore scroll position and cursor when navigating back
* virtualized list rendering (so you can navigate to a directory with 10,000 files without stuttering)

And comes with some nice edge cases. For example, if you select row 30, then scroll it out of view with the mouse wheel, and then press enter, it needs to start playing row 30, even though row 30 no longer exists in the component tree/DOM (due to virtualization). And then row 30 should be scrolled back into view so you can see what just happened.

The browser's built-in focus mechanism does not work with virtualization. When you scroll a focused element away in a virtualized view, it is removed from the DOM, and the browser resets focus to the body element. So the browser's focus must be fixed to the virtual scroll element, and a synthetic focus must be maintained within it.