Loosely match "now playing" song
Match by href instead of by currContext (referential equality) + currIdx.

Problem: the decorated "songIdx" was getting out of sync on the favorites when a favorite was added or removed. I decided to remove the idx from items in favorites; it's not needed. The idx is still decorated for items in browse listings in order to index files separately from directories (and matches the play context indexes).